requirements.txt
streamlit
streamlit-extras
google-generativeai
python-dotenv
Pillow
